SERVER-98080 Parallelize python test fixture setup (#30047)
During setup of sharded cluster python fixture, we spawn several
mongod/mongos processes, and [we do this
serially.](https://github.com/mongodb/mongo/blob/e103eaed0855666ec274105f3f4622ed748326e2/buildscripts/resmokelib/testing/fixtures/shardedcluster.py#L178-L192)
When the test cluster contains many shards and several replicaset nodes,
this can take very long time.

To speed the fixture setup phase, we can spawn the different processes
in the cluster in parallel. This will contribute to reduce the time to
wait for the fixture setup to complete and thus reduce considerably the
time it takes to run a JS test.

These are the numbers that characterize the setup of a cluster made of 1
mongos, 1 config server and 2 shards with 3 nodes each:

    [Current] Serial setup: ~25 secs
    [Optimized] Parallel setup: ~9.5 secs