SERVER-98341 Fix compile_upload_benchmarks task failure (#30217)
A change from the previously merged SERVER-98092 broke the
compile_upload_benchmarks task in Evergreen for the [A,UB]SAN Enterprise
RHEL 8 DEBUG (Benchmarks) variant.

A hot BF got opened because the task is failing for all new runs on the
`[A,UB]SAN Enterprise RHEL 8 DEBUG (Benchmarks)` variant.

A workaround is to revert part of the changes from SERVER-98092 until
the issue is better understood.
This PR converts the following libraries from Bazel back to SCons:

* `query_bm_constants`
* `query_bm_fixture`