SERVER-95523 upsert does not handle DuplicateKey retry correctly (#28489)
When updates and findAndModify run with upsert:true, they can retry in
certain cases when the insert hits a duplicate key error. The assumption
made is that another query inserted the conflicting document
concurrently - otherwise the upsert would've modified the conflicting
document.

This assumption can be broken when the unique index does not contain
every document in the collection. When the update sends a document into
the unique constraint, a conflicting document may have existed prior to
the upsert query.

The bug was that this could retry indefinitely. To fix, cap the number
of retries.