SERVER-98322: Improve field path validation error messages (#30219)
When parsing for `ExpressionFieldPath` and `ExpressionObject`, we
provide an error message encouraging users to use `$getField` or
`$setField` for field names with dots and dollars. That error message is
now no longer given for other use cases of field path validation (like
`$facet` and `$rankFusion` pipeline names) where that alternative is not
an option.