SERVER-98376 Fix hardcoded useE2EAPI and add missing boolean and null conversion (#30253)
This change only affects benchmark and unit testing. This commit fixes
hardcoded useE2EAPI and adds missing boolean and null conversion. Also,
it extends the string test cases and increases the generated string
length up to 32 in order to show the significance of string length on
performance. It removes kZipfian to save benchmark running time as that
showed little impact on perf.