SERVER-98398 Include enableTestCommands in config fuzzer generated mongos.conf (#30297)
A recent change added a failpoint as a target for the config fuzzer to
fuzz for mongos. However, it didn't add a setParameter for
enableTestCommands, which can lead to the configuration being rejected.
This isn't a problem in the config fuzzer tasks as those pass
enableTestCommands to mongos on the command line. However, the jepsen
config fuzzer tasks don't do that, so they started failing.

This is the first time a failpoint was a target for fuzzed mongos
configs, causing the error to arise now. We already set
enableTestCommands in fuzzed configs for mongod.