SERVER-96872: Implement propagate schema for $lookup w/ CSFLE (#29966)
This change introduces EncryptionSchemaEncryptedObjectArrayNode,
representing the encrypted array
resulting from a lookup stage. Encrypted arrays are not permitted in
encryption schemas, which is
enforced by the initial schema parsing code. However, the result of a
lookup with an encrypted
foreign collection is an encrypted array, whose contents should not be
usable in stages following
the lookup. This new node restricts the usage of the encrypted array
(and its contents) in
subsequent stages, but also allows for the array to be used if it is
unwound by a subsequent unwind.