SERVER-81967 Rely only on `getDatabaseVersion` to get the cached database version from MongoS (#30371)
The `getShardVersion` command retrieves database metadata when a
database name is provided as input. As part of the MongoS Interface
Parity project, routers now support the more explicit
`getDatabaseVersion` command.

In this context, existing tests that previously used the
`getShardVersion` command on the router to obtain database metadata have
been updated to utilize the `getDatabaseVersion` command instead.