SERVER-98366 Mark query_settings_plan_cache as incompatible with bala… (#30243)
Created from https://jira.mongodb.org/browse/BF-35935.

The issue appears to be that the SBE plan cache key includes information
about [the
collection](https://github.com/10gen/mongo/blob/acfa58cc693937fe0f5c514940a232134054c724/src/mongo/db/query/plan_cache/sbe_plan_cache.h#L143).
If a collection is moved, from the logs, a DDL operation is triggered,
which changes this value. The test ends up comparing a plan cache key
that was generated via explain() before the move, with a plan cache key
generated after. The solution here is to just ban the test from this
suite by marking it as assumes_balancer_off, as is done for other tests
in jstests/core/query/plan_cache. (E.g.
[here](https://github.com/10gen/mongo/blob/acfa58cc693937fe0f5c514940a232134054c724/jstests/core/query/plan_cache/plan_cache_distinct.js#L13-L14)).

More complete explanation from slack conversation:

The test essentially has a race condition when run under that suite. It
runs a command which writes to the plan cache, and then runs a separate
command, an explain, to compute the plan cache key. It then makes an
assertion which assumes the key it computed with explain was the same as
the key written. When the test suite runs a DDL operation on the
collection in the background, and it happens to be in between the two
commands, the keys used for those two operations can be different.
Here's roughly what happens:
```
a) Test runs query which writes to to the plan cache using key <X>
b) (in the background) the suite moves the collection
c) Test runs explain to compute the plan cache key. Due to (b) the key is actually computed differently, and key <Y> is returned.
d) Test attempts to lookup cache entry <X> using key <Y> and fails
```

In normal operation this would be fine, you would just get a cache miss.
But since the test is assuming that the cache is not touched for the
duration between (a) and (d), it fails.