SERVER-97726: Prevent directory deletion racing against ident creation (#29972)
# Issue

[SERVER-97726](https://jira.mongodb.org/browse/SERVER-97726)

# Description

Its possible for a directory to be removed between when the kv engine
creates it and when wiredtiger depends on it existing. To prevent this,
we use a mutex, preventing creations and deletions of directories from
occurring concurrently.

Serializing deletions is not ideal, but the complexity of tracking
per-directory mutexes is somewhat cumbersome and probably not worth it.
The vast majority of "drop" calls won't ever hold the mutex.

We also opportunistically fix an issue where only 1-deep directories
would be deleted, now empty directories will be deleted recursively.