SERVER-96875 Implement schema propagation for lookup w/ FLE2 (#30349)
This change handles propagating the encryption schema for $lookup and
FLE2 schemas. For the FLE2 case, we don't support using encrypted fields
coming from the right hand side of the $lookup, even after an unwind of
the array (unlike CSFLE). To achieve this, we don't reset the
_isWithinEncryptedArrayContext flag on encrypted nodes after an unwind
for FLE2 schemas. This prevents referencing the encrypted nodes after
the unwind.