SERVER-93788 Fix PBT helper function to provide correct number of generated query shapes (#30452)
Currently some PBTs are silently not running because they
[reference](https://github.com/10gen/mongo/blob/d7a247e1f4bb39386fb0063f62ce37b6532d5d34/jstests/libs/property_test_helpers/query_properties.js#L28)
`numQueryShapes` and the test runner names this field `numQueries`.
Javascript silently allows this comparison and returns false, so two of
the PBTs don't run any queries.

In the future, we'll have the tests print out stats about their
execution to prevent this, but that should be a separate ticket.