SERVER-97836 Add new tokens for QE text search (#30444)
Most of the diffs in this patch come from the FLEXXXTokenGenerator
classes being removed. Instead of the generators, we now use
`FLEXXXToken::deriveFrom()` to generate new tokens. Tokens based on the
root key now take a `const FLEIndexKey &` as an argument to
`deriveFrom()`.

Besides that, we add all 24 new tokens from the QE text search spec, and
update the token test to ensure these tokens are being generated
correctly.