SERVER-96793: Add "scoreDetails" field to DocumentMetadataFields (#30396)
This patch adds the `scoreDetails` as a valid metadata field.
`scoreDetails` must be a `Document` and, by default, we will also
populate the `score` metadata field by retrieving its "value" field.
This will be important in the future when implementing scoreDetails for
$score.

We handle this slightly differently for `searchScoreDetails`. When
setting `searchScoreDetails`, we will also set `scoreDetails` (in line
with the `searchScore` <> `score` relationship), but we do _not_
populate `score` since search will separately populate its own score
values.