SERVER-98679 Add --nobest flag to dns docker container (#30542)
Adds a flag `--nobest ` were if for any reason the latest version of a
dependency cant be obtained it will fall back to an older version.

This would have prevented
[BF-36027](https://jira.mongodb.org/browse/BF-36027) where it couldn't
pull the latest version of a dependency and it kept trying for 2 hours.

Also fixes a bug in the `retry` where the counter wasn't incremented
becoming an infinite loop.