SERVER-97864 Support sampling of IndexScanNode (#30346)
This patch invokes the sampling estimator for IndexScanNode to estimate
the number of keys it will scan, and the number of documents
(deduplicated RIDs and an optional filter) it will return. This patch
also attempts to recognize Fetch -> IndexScan plans and sample the
cardinality of the fetch node the same way the number of documents for
an index scan with a filter be estimated. This allows for multiple plans
with different access paths to have the same estimate.