SERVER-97308 Skip creating .a intermediate files when static linking on linux (#29399)
Add start_end_lib_feature flag to Bazel so our program files get built
against the .o files when static linking on linux rather than the .a
files. This saves us from having to create the intermediate .a files.
Added a skip to extract_debug_info because we don't extract debug info
from the .a files, and we don't want to create them by taking a
dependency on them.

---------

Co-authored-by: Daniel Moody <dmoody256@gmail.com>