SERVER-98364 Annotate individual queries with feature info in the case of failure (#30291)
This PR pretty-prints a list of features associated with a failing query
id and query when the --extractFeatures flag is on.
Since features are tracked per query, they are no longer tracked in
aggregate as the former should help with more directed debugging help.

<img width="895" alt="Screenshot 2024-12-16 at 11 33 17 AM"
src="https://github.com/user-attachments/assets/08f52ce4-6df3-4228-9182-cd9c8411abea"
/>

Note that this depends on
[DEVPROD-13531](https://github.com/10gen/feature-extractor/pull/36).