SERVER-95993 Suppress -Winterference-size (#30577)
We previously had added the flag -Wno-interference-size in
`mongo_cc_toolchain_config.bzl` - making it apply any time we compile
C++ code with the mongo toolchain. However, it's not understood by clang
in the v4 toolchain, so we also need -Wno-unknown-warning-option so that
compiler will ignore it. We set that flag in `mongo_src_rules.bzl` so it
applies to any `mongo_cc_*` target.

This mismatch led to a BF in the gRPC variant because we'd build
protobuf using our toolchain and therefore -Wno-interference-size, but
internally it has `cc_library` rules rather than `mongo_cc_library`
rules (e.g., the rule
@@com_google_protobuf//src/google/protobuf/compiler/cpp:cpp is a
`cc_library`).

This change moves -Wno-unknown-warning-option to be a toolchain-level
option and adds back the -Wno-interference-size size option.

---------

Co-authored-by: James Bronsted <32047428+jpbronsted@users.noreply.github.com>