SERVER-98310 Create interface for shared libraries linking static libraries in Bazel (#30191)
Ports over top-level shared archive compilation and linking to Bazel.

Shared archives are stand-alone shared libraries that statically link in
all of their transitive dependencies and can be more easily distributed
(you don't need to ship all of the transitive shared libraries it
depends on). You also get rid of a significant portion of the runtime
performance loss from using shared libraries. We use this for
stitch_support and mongo_crypt_v1, which we ship to customers to link
against in their product.

Bazel cc_shared_library's implementation ignores static linker inputs
that are owned by a different target than the cc_shared_library itself.
Since we need to transitively collect the list of static linker inputs
for the shared archive implementation, we need to transtiively remap
ownership of each linker input to make Bazel actually link them in.

Ex:
**mongo_crypt_v1.so**
-> libserver_base.a (linker input owned by mongo_crypt_v1.so)
-> -> libbase.a (linker input owned by libserver_base.a, will be ignored
normally and not linked into the final .so)

Since we effectively want to flatten this down into:
**mongo_crypt_v1.so**
-> libserver_base.a, libbase.a (both owned by mongo_crypt_v1.so)

We need to the remap libbase.a linker input to be owned by
mongo_crypt_v1.so

This also creates a top level install target for shared libraries since
they need to be installed into a specific aib package

Also reduces the number of full compilations done in the required build
by another 2