SERVER-98442 Use existing tie breaking for distinct scan multiplanning (#30317)
Removes the tassert & direct bonus for any distinct scan plan, and
instead uses the multiplanner tie breaking heuristics to apply a bonus
for distinct scan plans only in case of a tie. This also allows distinct
scan plan score ties to be broken based on the same heuristics that
index scans use.

---------

Co-authored-by: Peter Volk <77723508+HCSPete@users.noreply.github.com>
Co-authored-by: David Percy <david.percy@mongodb.com>