SERVER-95135 Enforce that originalSpec has same collation as outer spec when creating index (#30186)
Restricts the behavior of input `originalSpec` by ensuring that the
collation matches the index collation:

Pass Cases:
1. No collation specified on both the `originalSpec` collation and the
index collation
2. Collation specified on both `originalSpec` collation and index
collation and they match
3. Collation specified on `originalSpec` collation only, matches default
collection collation
4. Collation specified on index collation only, matches default
collection collation
added a check in the helper `verifyAndDropIndexes` that ensures that the
originalSpec.collation and the indexSpec.collation will be equal in the
pass cases.

No-Pass Cases:
1. Input `originalSpec` collation and index collation, but they both
don’t match
2. Input collation specified on `originalSpec` (doesn’t match default
collection collation), none on index
3. Input collation specified on index (doesn’t match default collection
collation), none on `originalSpec`

These cases are all tested. Let me know if there are any cases that
shouldn’t be a pass case or any cases that shouldn’t be non-pass cases.
Note: I have a commit version where the uasserts in
`validateOriginalSpecCollation` are within the
`validateIndexSpecCollation` and we pass in a `boost::optional` for the
`indexSpecCollation` -- I don't have strong opinions about if these
should be two separate functions or the same.

---------

Co-authored-by: Shin Yee Tan <shinyee.tan@mongodb.com>