SERVER-98738 Add way to change default config in bazel (#30740)
It can be annoying to always have to pass --config on the command line
when you want to stay on one config. Also accidentally forgetting to
pass --config will cause you to discard the analysis cache. Add an
expected way for users to set the default config so they can stay in it
while building.