SERVER-93729 Implement the Chunk-based sampling CE approach (#30523)
This patch implements the chunk-based sampling CE approach. This
sampling method generates a sample by randomly picking starting point
for each chunk of documents, then sequentially scanning a specified
number(this number is the size of a chunk) of documents. This patch is
also introducing a new internal query knob for specifying the number of
chunks to draw the sample.