SERVER-98477: Implement core logic for insertBatch (#30379)
This is an attempt to define the core logic for insertBatch. The key
parts are:

timeseries_write_util::insertBatchOfMeasurements: this takes in a vector
of measurements and returns a vector of WriteBatches for those
measurements. I had this in mind as the top-level interface for
insertBatch, with this being called in timeseries_write_ops when we
decide how to hook it into the unordered inserts write path. One
assumption that this currently has is that all of the inputted
measurements currently belong to the same time-series - that is, that
they have the same metafield value. I think that one of the next steps
would be to have this function handle splitting up the vector of
inserted measurements into time-series (sections with the same
metaField), although that could also happen at a level higher than
this. Two other considerations that I'm open to thoughts on : Do we want
to have one writeBatch for all of the measurements in a single
time-series, instead of the vector of them as is currently defined
here? One optimization that we can do is to sort the measurements that
we are inserting. This would reduce the chances of time-related
rollovers. If we do this, do we need to preserve the notion of what the
original ordering was? The same applies for when we take a vector of
measurements that could be from different time-series and split them up
by metaFields: do we need to keep track of what the original ordering
was?

bucket_catalog::insertBatch: this is called into by
insertBatchOfMeasurements, and the core logic that will be handled here
will be handling finding buckets to insert measurements into. In this
current, simple implementation, it does this by allocating a new bucket
to begin with, and then inserting into it through
bucket_catalog::insertBatchHelper. It handles the helper method
returning a rollover reason (time-related) by allocating a new bucket,
and then continuing to insert into it. Future work will incorporate it
going through the usual stream of :
 A) Checking for a bucket in-memory to use 
 B) Checking for a bucket to reopen 
 C) Then allocating a new bucket to insert into
 When we decide to try to find a bucket on disk to reopen and insert
into, this will involve releasing the stripe lock temporarily.
 
bucket_catalog::insertBatchHelper: this implements the core logic where,
given a bucket to insert into and measurements to insert, it will fill
up the current bucket to the brim. If while inserting a measurement into
the provided bucket, the bucket is rolled over for a hard reason (i.e,
count, size, etc), it will allocate a new bucket while rolling over the
provided one, and change the inputted bucket pointer to reflect this. If
the rollover reason is time-related, it returns the rollover reason and
pops out of the loop, keeping track of progress by incrementing the
currentPosition counter. Upon inserting all inputted measurements, it
will return a std::monostate.

One change outside of these newly defined functions is to have
bucket_catalog_internal::insertIntoBucket modify the inputted
existingBucket pointer upon rolling the existingBucket over for a
time-related reason. This is because bucket_catalog::insertBatch, upon
receiving a rollover reason from the helper, in the process of handling
the rollover reason has to know that it is dealing with the bucket that
we are currently filling up. If we gave the helper function Bucket A to
insert into, and it chugged along filling up Bucket A entirely, then
rolling over and filling up Bucket B and Bucket C entirely, and then
while filling up Bucket D it encountered a time-related rollover reason
and popped out, we have to know that while dealing with the rollover we
can act on Bucket D. insertIntoBucket does not currently return this
value, and I thought that this was cleaner than adding another return
value to the function in general; if this seems dangerous I am open to
changing the design.