SERVER-98625 Fail ASSERT_STRING_SEARCH_REGEX if the regex is invalid (#30775)
Currently, ASSERT_STRING_SEARCH_REGEX passes if given an invalid regex.
This makes it easy to accidentally write an invalid regex (e.g., by
having unbalanced brackets) and think your test checks for a match when
it fact it passes vacuously. This also makes DEATH_TEST_REGEX have the
same problem - a death test with an invalid regex passes as long as the
child process dies with a nonzero exit code, independent of the output.

This change makes ASSERT_STRING_SEARCH_REGEX fail if the regex is
invalid and fixes one death test that was using an invalid regex.