SERVER-97711 Avoid initialization of RoutingInformationCache under FeatureFlagDualCatalogCache (#30129)
When `FeatureFlagDualCatalogCache` is enabled, the shards and config
server use a separate catalog cache for filtering and routing purposes
(shard and router role). The config server has an additional catalog
cache, `RoutingInformationCache`, that is also used for some routing
tasks specific to the config server. This patch avoids initializing it
under `FeatureFlagDualCatalogCache` since it is redundant with the
routing catalog cache.

---------

Co-authored-by: Pol Pinol Castuera <pol.pinol@mongodb.com>