SERVER-99023 Explicit Cardinality estimation for every QuerySolutionNode (#30938)
This patch achieves the following:
* Add CE for STAGE_EOF. This fixes the assertions thrown by empty $in
and $all predicates, which get optimized to EOF plans.
* Remove the default case in the switch statement which dispatches the
CE calculation. This gives us compiler-time safety that we handle every
possible stage. This patch marks stages which we still need to implement
with an unimplemented tassert and the ones which won't reach CBR with a
unreachable tassert.