SERVER-99038 Return resolved pipeline for $lookup serialize except for explain (#30884)
Returns `_resolvedIntrospectionPipeline` for $lookup serialize in order
to allow FLE2 to work properly in a sharded environment.

When sending the aggregate pipeline to a shard, the opts parameter is
not included in the serialization function. For FLE2, we need to have
the `_resolvedInstrospectionPipeline` returned in this scenario as it
contains rewritten information required for correct query execution.
Because of this, we modify the $lookup serialize function to return
`_resolvedInstrospectionPipeline` when opts is not given. We exclude the
explain case (to be done in SERVER-81802), as additional work is
required to address that scenario.