SERVER-99163: Remove coordinateCommitReturnImmediatelyAfterPersistingDecision from change_streams* suites (#31000)
Removing coordinateCommitReturnImmediatelyAfterPersistingDecision=true
from the all change stream suites configurations as it means 2PC
transactions return as soon as decision is persisted which is before
commitTransaction has been run on shards. This violates read-own-writes
when mixing reads and writes inside and outside of transactions. This is
why the production value of the server parameter is false