SERVER-98996 Make mirror_reads.js less racy (#30853)
This commit renames the mirrored reads `pending` stat to `scheduled` and
redefines the `pending` stat to refer to commands that will be mirrored
but have not scheduled yet. This is important because the mirror_reads
jstest may think that mirroring is completed if some reads are pending
but not yet scheduled (using the new definitions for `pending` and
`scheduled`). As such, the mirror_reads jstest is now better at telling
when mirroring is completed.