SERVER-99071 Simplify Sorter checksum calculation  (#30904)
Sorter fed data into the checksum calculator in two different ways: for
v1 each specific range of bytes consumed by deserialization was fed in,
and for v2 each block read from disk was fed in. This was possibly a
minor optimization for limited/top-k sorters which would not always
consume the entire block, but v1 checksums are now only used for
resumable index builds which always consume the entire input. As such
this code can now be simplified to always work in the v2 manner (which
produces the same checksum values as the old behavior for v1).

This also enables moving checking the checksum to the file read step
rather than closeSource(), which allows removing closeSource() entirely
(and openSource(), which has been unused for a while).