SERVER-56865 Allow explodeForSort to use index with multikey field following sort fields (#29640)
Currently the explodeForSort optimization will attempt to turn an index
scan over a union of point intervals into a set of point interval index
scans which can be merge sorted to satisfy a sort. The currently logic
to avoid this optimization if there is a multikey field in the index
after the exploded field. This patch looses this restriction to enable
the optimization to work in more cases; specifically, if the multikey
field is after the sort fields, then the optimization can still work.