SERVER-98374 Expose and use catalog information in CBR (#30901)
This patch plumbs the CollectionInfo struct to the CardinalityEstimator,
which gives us the ability to use multikey information when making
estimation decisions.

This patch also fixes the estimation of conjunctions over the same
mulitkey field using histograms. Before, we would naively assume that we
could always intersect intervals that we generated from predicates, but
now we only do so when the field is non-multikey.