SERVER-29967: Move filter before dedup in index scan (#30764)
Part 1 of SERVER-29967

## What does this PR do?

Suppose we are doing an index scan over a multi-key index.

On master, IXSCAN does the following:
1. First it tries to insert a recordId to the _recordIdDeduplicator,
which simultaneously checks whether or not the recordId is already in
the duplicator.
2. Then, if the recordId has not been seen until now, it runs the filter
against the key.

In other words, IXSCAN only looks at the first element of an array when
determining whether or not the recordId corresponding to a key is the
recordId of a document that can potentially be part of the result set.
In practice, this means that in order to return correct results, our
plans needlessly wait until the parent stage of a IXSCAN to apply some
filters that only reference key fields.

On this branch, IXSCAN instead:
1. Checks if the recordId exists in the _recordIdDeduplicator without
adding it.
2. Then, if the recordId has _not_ been seen before, it runs the filter
against the key.
3. If the filter returns true, the recordId is added to the
_recordIdDeduplicator

In other words, IXSCAN evaluates the filter against the elements of an
array one by one until it finds one that matches. This means filters can
run during the IXSCAN stage even when there is a multi-key index.

That said, until query plans are changed in a subsequent PR, filters
will not actually run during the IXSCAN stage when the index is
multi-key.

## Performance implications

With each recordId that passes the filter, this PR would add one
additional operation to check if the recordId exists in the
_recordIdDeduplicator. This is because the act of inserting something
into the _recordIdDeduplicator includes an existence check. As I don't
think that the existence check is likely to be very expensive, I think
this is better than evaluating a filter against keys corresponding to a
recordId that IXSCAN already returned.

If there is no filter, I insert into the recordId into the
_recordIdDeduplicator at the same time that I check for existence, but
this might be an over-optimization.

## Testing

Since filters will not run when there is a multi-key index scan until
plans are changed, I added a unit test to be able to see the new
deduplication behavior.