SERVER-99226 Mark some PlanStage and PlanExecutor methods as const (#31044)
The following methods should be declared const as they do not modify any
state of the queried objects:

* `PlanStage::isEOF()`
* `PlanExecutor::isEOF()`
* `PlanExecutor::getKillStatus()`

This should not change behavior or performance in any way, but
strengthens the API contracts of these methods. Callers would not expect
these methods to modify the state anyway. By marking them as explicitly
const, we can prevent them from unintentionally modifying state in the
future.