SERVER-98771 Fix use-after-move in ABT Constant::array helper (#31074)
This function invoked std::forward on the variadic parameter in the fold
expression twice so that it could invoke Array::push_back(TypeTags,
Value).

This patch changes the fold expression to invoke
Array::push_back(std::pair<TypeTags, Value>) and as a result only
perform the std::forward on elements once.