SERVER-99404 improve wrapper module install detection (#31181)
This change makes the wrapper look anywhere for installed modules, as
the modules are not config specific so really we can import anyone that
is available. This fixes the cases where the wrapper would end up
looking in what ever config you happen to build in last, and then
switching the config on you while you are building.

This also moves the module code into common python code simplfying the
platform wrapper scripts and making the code a bit easier to read.