SERVER-80943: Add a test for change stream output for time-series shardCollection operation (#29967)
[SERVER-80943](https://jira.mongodb.org/browse/SERVER-80943)

Adds a test to generate and validate the structure of the
'shardCollection' change stream event. This ticket came out of the bug
identified in
[SERVER-79974](https://jira.mongodb.org/browse/SERVER-79974) where the
name of collection in the 'shardCollection' was improperly changed and
went unnoticed until it caused a problem.

This jstest attempt to validate as many fields/variations as seems
necessary. Please feel free to suggest additional variations upon review
(I am not a time-series collection expert).