SERVER-99396: Handle case where test file is in cwd (#31210)
If parent path is empty then we are in the same directory as the test
file and we can proceed. However, if parent path is non-empty then we
should guarantee existence of the parent directory.

This PR resolves the crash that occurred when running `mongotest` and
passing in a test that was in the cwd e.g.
1. `mongotest -t a.test --drop --load --mode run` would fail (PR fixes
this case)
2. `mongotest -t tests/a.test --drop --load --mode run` would succeed