SERVER-99340 Simplify WiredTiger bulk index insertion (#31129)
Inserting into Id indexes does not need to build a new buffer with the
record id and typebits because key_string::Value already stores data in
exactly the required format.

key_string::Builder::appendTypeBits() explicitly appended a zero to
represent all zeroes, and then that zero was stripped when inserting it
into the index. Everything which uses TypeBits supports representing all
zeroes as an empty buffer, so we can skip a step.