SERVER-98554: Fix ASAN stack use-after-free failures (#30507)
This PR fixes the stack use-after-free issue in the Kafka resolver
callback (pointer ownership is handed to rdkafka's resolver code, but
it's returned via .get(), and then the unique_ptr is deleted after when
the function completes). Updated, and repaired the tests to ensure they
call the cleanup code as rdkafka does as well.