SERVER-98998 Prevent race in new_primary_prevents_stale_majority_reads_on_stepup where we use the new primary entry ts to update the snapshot instead of the target write ts (#31263)
To minimize the race where the new primary uses the updatePosition from
writing/applying the new primary entry instead of the updatePosition of
the target write, we manually call awaitReplication after stepUp and
wait for the primary's lastWritten opTime to replicate to the secondary.
In addition, the primary should also update its perception of the
secondary's lastWritten opTime as this implies it successfully received
the new primary updatePosition.