SERVER-99130: Limit query_tester_generated_test_* to Query Patch Alias (Try SBE Engine) with waterfall support (#31240)
- The `query_correctness_generated_test_*` (previously
`query_tester_generated_test_*`) tasks should only run on the Query
Patch Alias (Try SBE Engine) as that is the config we end up shipping to
customers.
- The `experimental` tag lets us have more control over which variants
it runs on when we eventually enable it on the waterfall. This avoids
running it on the "significant development variants". More info
[here](https://github.com/10gen/mongo/blob/master/docs/evergreen-testing/yaml_configuration/task_selection_tags.md)
- I created a new `query_tester` evergreen task tag to identify these
tasks
- I copied over the contents of the `query-alias-task-list` YAML anchor
and added `.query_tester` as an acceptable tag while disallowing it in
the other 2 Query Patch Alias variants.