SERVER-99486 Change QueryTester test number expectations to only strictly-increasing (#31282)
This PR makes it so that `testNum`s only need to be strictly increasing,
not necessarily in consecutive order. The `-r` flag will still allow
running tests in a range of `testNum`s, and the `-n` flag will still
reference the `testNum` as well. If a query/test is unlabeled, it will
automatically be assigned the sequential `testNum` directly after the
previous `testNum` in `--mode run --out result`.