SERVER-99466: Coveralls reports show different commit than the PR (#31235)
Setting environment variables based on the `${github_commit}` expansion:
* `COVERALLS_GIT_COMMIT` for unittests via the coveralls reporter
* `TRAVIS_COMMIT` for jstests via the gcovr collector/reporter

These are captured in the data that is sent to Coveralls to render
results and comparisons. Before, the tools were just sniffing out what
git commit it was running from, which was a little misleading.

Now coverage stemming from PRs have the correct commit sha (instead of
some random one from master) with the correct version of the files so
that line counts and file diffs accurately show users' latest changes.

For non-PR workflows, this commit is still mismatched, pointing to
something else on master due to Evergreen git-behaviors; I am engaging
on how this can be applied to those as well, but outside of this PR
(related to DEVPROD-13619).

Also refactored the logic on the file at large to be more readable, and
deleted some dead code since Bazel has fully come aboard. It might help
reviewers to look at the initial prefactoring step in
https://github.com/10gen/mongo/pull/31235/commits/fbca0162bc89b28da897372ac20fa62e4c1b0b59
if the diff is too much all at once.