SERVER-99597 Increase query benchmark timeout by one hour (#31330)
Similar to [SERVER-98746](https://jira.mongodb.org/browse/SERVER-98746),
we saw another increase in the query benchmarks runtime, where it now
times out very often. We plan on removing these benchmarks from Windows,
but in the meantime while we wait for that to be finalized, we'll
increase the timeout again.