SERVER-99096 Allow histogram estimation in cases when only prefix of index is used and unbounded suffix does not have histograms (#31384)
This patch removes the constraint that all paths of an index must have a
histogram in order to estimate in index scan, even if only a prefix of
the index is used (where unused fields are unbounded). For example, a
query on 'a' with index (a, b) and histogram only on 'a'.