SERVER-99468 Introduce key_string::View (#31242)
This is a non-owning view into a buffer created by key_string::Builder.
It is a drop-in replacement for key_string::Value in places that do not
need to take ownership of the buffer. As an initial step this is used
only in some of the SortedDataInterface functions where it was trivial
to adopt. The Value -> View conversion is currently implicit to avoid
having to touch every caller, but may make sense to be explicit once it
has been propagated outward.