SERVER-99523 Do not restoreState() when catch temporary unavailable exception from storage (#31343)
PlanYieldPolicy::yieldOrInterrupt() should not call a full
restoreState() in the catch StorageUnavailableException block but
instead just restore the contents of '_yieldable'. saveState() is
already called again at the start of each retry and the code has been
designed to allow multiple saveState() calls in a row, with no
restoreState() in between, for at least ten years. Calling
restoreState() in a retry can trigger invariant() failures if the state
had not been completely saved due to an exception. restoreState() itself
can also throw, so it must not be called in the catch block to prevent a
new StorageUnavailableException from causing a failure instead of
another retry.