SERVER-99661 Sync compile_all_but_not_unittests cache with other tasks (#31412)
This task currently has its own compilation cache due to disabling
debug-symbols at the compilation stage. This results in 3 full rebuilds,
putting significant load on the RBE cluster. Re-use the compilation
cache from other tasks.

The main reason for removing debug symbols is that the IO load on the
linker step is significantly higher, but we can achieve the same result
by keeping the debug symbols in the compilation output and instead
stripping them at the linker step. This way we will avoid invalidating
the compile actions, but keep the same performance win since linker
actions are local and always uncached.